oracle_instance: added SCN headroom to service description
The current SCN and the usage of the SCN headroom for scheme 1 has been added
to the service description. SCNs (System Change Numbers) are unique over the
whole live of a database. The maximum SCN depends on the database version and
the current time to save enough SCNs for the future. Running out of SCNs is a
really rare situation. The headroom is the area between current SCN and maximum
allowed SCN for the current time.

There are no parameters for the headroom, because it highly depends on the
Database version and installed patches.

Please read Oracle Support Note 1376995.1 for more details about SCN headroom
and affected Database versions.

Change-Id: I2bfe2de13f66caba1437a5783fdc490f60f4a9a2